+++ server/vulnerabilities/msrc/msrc_api.go
@@ -1,4 +1,4 @@
-package io
+package msrc
 
 import (
 	"errors"
