+++ server/vulnerabilities/msrc/msrc_api_test.go
@@ -1,4 +1,4 @@
-package io
+package msrc
 
 import (
 	"net/http"
